Learned pair-wise classifiers
These are the neural network classifiers that were learned by the code for the case study 2 in the paper. If you retrain the neural networks, you might get other results since the training is not always converging. Some hyperparameters surely need to be tuned; but we don't have time for this; and we don't have the compute power for doing a good job about this; so it's hopeless. 

But at least it shows that the code can work and give something non-trivial statistically.